Remove network from user and adapt filter mixin
